Fix bug introduced in this branch custom locations
The commit that allows loading datasets from custom locations refactored a lot
of code and made the datasets.utils module more functional.
However, we left a stray self in the download function, which was not caught by
the tests, because the tests do not download any resources (AFAIK).

This should fix that.